requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
joblib==1.4.2
pandas==2.2.3
requests==2.32.3
paho-mqtt==2.1.0

numpy>=2.0.0,<3
scipy>=1.13.0,<2
scikit-learn>=1.4.2,<2
